ADD: some details in sand

+++ Coursework/coursework/Primitives/Model/model.cpp
@@ -77,7 +77,7 @@ void Model::add_side(std::initializer_list<size_t> ind_arr, QRgb color)
     _add_side(new_side, color);
 }
 
-void Model::add_side(vector<int> ind_arr, QRgb color)
+void Model::add_side(vector<size_t> ind_arr, QRgb color)
 {
     vector<shared_ptr<Vertex>> new_side;
